+++ slick/src/main/scala/pl/iterators/kebs/Kebs.scala
@@ -1,7 +1,7 @@
 package pl.iterators.kebs
 
 import pl.iterators.kebs.macros.CaseClass1Rep
-import slick.ast.BaseTypedType
+import slick.ast.{BaseTypedType, NumericTypedType}
 import slick.lifted._
 
 import scala.language.implicitConversions
@@ -11,10 +11,24 @@ trait KebsColumnExtensionMethods {
     new StringColumnExtensionMethods[CC](rep)
   implicit def stringValueOptionColumnExt[CC](rep: Rep[Option[CC]])(
       implicit ev: CaseClass1Rep[CC, String]): StringColumnExtensionMethods[Option[CC]] = new StringColumnExtensionMethods[Option[CC]](rep)
+  implicit def numericValueColumnExt[CC, B](rep: Rep[CC])(
+      implicit ev1: CaseClass1Rep[CC, B],
+      ev2: BaseTypedType[B] with NumericTypedType): BaseNumericColumnExtensionMethods[CC] = new BaseNumericColumnExtensionMethods[CC](rep)
+  implicit def numericValueOptionColumnExt[CC, B](rep: Rep[Option[CC]])(
+      implicit ev1: CaseClass1Rep[CC, B],
+      ev2: BaseTypedType[B] with NumericTypedType): OptionNumericColumnExtensionMethods[CC] =
+    new OptionNumericColumnExtensionMethods[CC](rep)
+  implicit def booleanValueColumnExt[CC](rep: Rep[CC])(implicit ev: CaseClass1Rep[CC, Boolean]): BooleanColumnExtensionMethods[CC] =
+    new BooleanColumnExtensionMethods[CC](rep)
+  implicit def booleanValueOptionColumnExt[CC](rep: Rep[Option[CC]])(
+      implicit ev: CaseClass1Rep[CC, Boolean]): BooleanColumnExtensionMethods[Option[CC]] =
+    new BooleanColumnExtensionMethods[Option[CC]](rep)
 
-  @inline implicit def getCCOptionMapper2TT[B1, B2: BaseTypedType, BR, CC](
+  @inline implicit def getCCOptionMapper2TT_1[B1, B2: BaseTypedType, BR, CC](
       implicit ev: CaseClass1Rep[CC, B1]): OptionMapper2[B1, B2, BR, CC, B2, BR] =
     OptionMapper2.plain.asInstanceOf[OptionMapper2[B1, B2, BR, CC, B2, BR]]
+  @inline implicit def getCCOptionMapper2TT_2[B1, B2, BR, CC](implicit ev: CaseClass1Rep[CC, B2]): OptionMapper2[CC, CC, BR, CC, B2, BR] =
+    OptionMapper2.plain.asInstanceOf[OptionMapper2[CC, CC, BR, CC, B2, BR]]
   @inline implicit def getCCOptionMapper2TO[B1, B2: BaseTypedType, BR, CC](
       implicit ev: CaseClass1Rep[CC, B1]): OptionMapper2[B1, B2, BR, CC, Option[B2], Option[BR]] =
     OptionMapper2.option.asInstanceOf[OptionMapper2[B1, B2, BR, CC, Option[B2], Option[BR]]]
@@ -41,7 +55,7 @@ trait Kebs extends KebsColumnExtensionMethods {
     })
 
   private class StringMapIsomorphism[A](comap: String => A)
-      extends Isomorphism[Map[String, A], Map[String, String]](_.mapValues(_.toString), _.mapValues(comap))
+      extends Isomorphism[Map[String, A], Map[String, String]](_.map(kv => kv._1 -> kv._2.toString), _.map(kv => kv._1 -> comap(kv._2)))
   implicit final val intMapValueColumnType: Isomorphism[Map[String, Int], Map[String, String]]   = new StringMapIsomorphism[Int](_.toInt)
   implicit final val longMapValueColumnType: Isomorphism[Map[String, Long], Map[String, String]] = new StringMapIsomorphism[Long](_.toLong)
   implicit final val boolMapValueColumnType: Isomorphism[Map[String, Boolean], Map[String, String]] =