+++ slick/src/test/scala/SlickPgTests.scala
@@ -1,9 +1,10 @@
 import java.util.UUID
 
 import com.github.tminglei.slickpg.ExPostgresProfile
-import org.scalatest.{FunSuite, Matchers}
+import org.scalatest.funsuite.AnyFunSuite
+import org.scalatest.matchers.should.Matchers
 
-class SlickPgTests extends FunSuite with Matchers {
+class SlickPgTests extends AnyFunSuite with Matchers {
   import pl.iterators.kebs.Kebs
   import slick.lifted.ProvenShape
 
@@ -41,20 +42,24 @@ class SlickPgTests extends FunSuite with Matchers {
 
   case class TestId(value: UUID)
   case class TestString(value: String)
-  case class Test(id: TestId, string: TestString)
+  case class TestNumeric(value: Int)
+  case class TestBool(value: Boolean)
+  case class Test(id: TestId, string: TestString, num: TestNumeric)
 
   class Tests(tag: BaseTable.Tag) extends BaseTable[Test](tag, "test") {
     import driver.api._
 
     def id     = column[TestId]("id")
     def string = column[TestString]("string")
+    def num    = column[TestNumeric]("num")
+    def flag   = column[TestBool]("flag")
 
-    override def * : ProvenShape[Test] = (id, string) <> ((Test.apply _).tupled, Test.unapply)
+    override def * : ProvenShape[Test] = (id, string, num) <> ((Test.apply _).tupled, Test.unapply)
   }
 
-  test("Column extension methods") {
+  test("String column extension methods") {
     """
-      |class TestRepository {
+      |class TestRepository1 {
       |  import PostgresDriver.api._
       |
       |  def toLowerCase(ilikeString: String): DBIOAction[Seq[TestString], NoStream, Effect.Read] =
@@ -65,5 +70,37 @@ class SlickPgTests extends FunSuite with Matchers {
       |  private val tests = TableQuery[Tests]
       |}
     """.stripMargin should compile
+  }
+
+  test("Numeric column extension methods") {
+    """
+      |class TestRepository2 {
+      |  import PostgresDriver.api._
+      |  def power: DBIOAction[Seq[TestNumeric], NoStream, Effect.Read] =
+      |    tests.map(t => t.num * t.num).result
+      |  def mult2: DBIOAction[Seq[TestNumeric], NoStream, Effect.Read] =
+      |    tests.map(_.num * 2).result
+      |  def lt0: DBIOAction[Seq[Boolean], NoStream, Effect.Read] =
+      |    tests.map(t => t.num <= 0).result
+      |  def abs: DBIOAction[Seq[TestNumeric], NoStream, Effect.Read] =
+      |    tests.map(t => t.num.abs).result
+      |
+      |  private val tests = TableQuery[Tests]
+      |}
+      """.stripMargin should compile
+  }
+
+  test("Boolean column extension methods") {
+    """
+      |class TestRepository2 {
+      |  import PostgresDriver.api._
+      |  def and: DBIOAction[Seq[Boolean], NoStream, Effect.Read] =
+      |    tests.map(t => t.flag && (t.num <= 0)).result
+      |  def or: DBIOAction[Seq[Boolean], NoStream, Effect.Read] =
+      |    tests.map(t => t.flag || (t.num > 0)).result
+      |
+      |  private val tests = TableQuery[Tests]
+      |}
+      """.stripMargin should compile
   }
 }